feat(landing): wire MCP tool detail to SDK client
Replaces the temporary buildLiveClient() that delegated to api.ts +
contracts.ts and polled GET /jobs/:id every 3 seconds with a real
@weavrn/sdk-backed implementation.

Dependency:
- adds "@weavrn/sdk": "file:../weavrn-sdk" so the landing consumes
  the locally-built SDK without going through a registry.
- adds a webpack fallback in next.config.mjs that stubs `fs` and
  `child_process` for the client bundle. The SDK re-exports a
  small git-helpers module from its main index that pulls those
  in for Node-side consumers; the landing only uses the
  browser-safe parts (WeavrnClient + tool modules), so stubbing
  the Node modules lets webpack tree-shake the git helpers
  cleanly.

src/app/tools/[provider]/[slug]/tool-detail.tsx (new):
- Holds every testable piece of the route — helpers,
  ToolDetailClient interface, ToolEvent shape, buildLiveClient,
  ToolDetailView component, and ToolDetailStatus state-machine
  union.
- Next.js page.tsx files only allow a fixed allowlist of named
  exports; pulling these out of page.tsx is the cleanest way to
  preserve test-import surface and unblock production builds.

src/app/tools/[provider]/[slug]/page.tsx (rewritten):
- Slim route shell only: fetches the tool via getTool, wires
  signer state from AppHeader, renders ToolDetailView. No more
  named exports — strictly compatible with Next 14's page export
  rules.

buildLiveClient (in tool-detail.tsx):
- Constructs a `new WeavrnClient({ signer, chainId })` once per
  invocation (chainId from NEXT_PUBLIC_CHAIN_ID, defaulting to
  84532).
- createJob/createEscrowETH/createEscrowERC20/linkJobEscrow/
  approveJob now delegate to WeavrnClient methods. The ETH path
  uses ethers.parseEther + the SDK's createEscrowETH; the ERC-20
  path uses approveEscrow + createEscrowERC20 (parseUnits with
  18 decimals as the default). Strategy address resolution maps
  the listing's strategy name onto WeavrnClient.strategies, with
  an explicit override allowed when the listing supplies one.
- linkJobEscrow uses the SDK's built-in indexer-lag retry loop
  (replaces the page-level 12-attempt loop).
- approveJob uses WeavrnClient.approveJob (release + complete
  with retry).
- subscribe replaces the 3-second polling loop with the SDK's
  SSE-based subscribe(). The bridge maps WeavrnEvent →
  ToolEvent and hydrates the job via apiGetJob on `job_delivered`
  so the page's existing `event.job?.deliverable_data` access
  pattern keeps working.
- getJob still uses the landing's session-based apiGetJob —
  users on this page already went through the wallet sign-in
  flow, so no signed-GET roundtrip is needed just to render the
  delivered output.

Page submit handler:
- Replaces the ETH-only `pricingIsETH` gate with a price-token
  detection: `ETH`/empty -> createEscrowETH; ERC-20 address
  (matches /^0x[a-fA-F0-9]{40}$/) -> createEscrowERC20; anything
  else -> readable error. ERC-20 support arrives as a natural
  consequence of the SDK rewiring (approveEscrow + ERC-20 create
  are SDK primitives).
- All other UX state machine steps preserved (submitting,
  funding, linking, waiting, delivered, releasing, done, failed)
  including the funding/linking deadline countdown and the
  release-and-complete button flow.

src/app/tools/[provider]/[slug]/live-client.test.ts (new):
- 13 vitest cases mocking @weavrn/sdk at the module boundary and
  asserting buildLiveClient passes the right parsed values
  through (parseEther, parseUnits, computed deadline, strategy
  address selection, override path), the subscribe bridge maps
  WeavrnEvent payloads onto ToolEvent and tears down cleanly
  even when the SDK's subscribe() resolves after the caller
  cancelled, and getJob continues to use the session-based
  apiGetJob.

src/app/tools/[provider]/[slug]/page.test.tsx:
- Updated import path: ToolDetailView/ToolDetailClient/ToolEvent
  now come from "./tool-detail" instead of "./page". All 13
  pre-existing test cases pass unchanged because the
  ToolDetailClient interface is preserved as the test seam.

Auth note: the page continues to use Bearer-session getJob for
the delivered-job fetch. Signed-GET via the SDK's getJobSigned
remains available for SDK-only consumers; landing users always
have a session and don't need it.

Test suite: 132 pass across 11 files. Typecheck clean.

Pre-existing build issue not addressed (out of scope): the
`output: "export"` config + dynamic [provider]/[slug] route
needs `generateStaticParams()` to build statically. This blocks
`npx next build` regardless of the SDK rewiring, and predates
this commit. The webpack compile phase now succeeds (the SDK
import tree resolves cleanly) — only the static-export
collection step fails on the missing generateStaticParams.
